New config structure Moved context options to own subconfig Added icons for every context menu config option
